Partially migrate to the new template
Partially migrate to the new template
