+++ .gitignore
@@ -3,3 +3,5 @@ README.md
 Icon?
 .sass-cache
 make
+min
+_sources
