--- admin/assets/include/footer.php
@@ -1,29 +1,0 @@
-<footer>
-  <div class="row">
-    <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12 footer-1">
-      <div class="container">
-        <p>
-          Some text
-        </p>
-      </div>
-    </div>
-    <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12 footer-2">
-      <div class="container">
-        <i class="pe-copyright"></i> Copyright 2016&nbsp;
-        <a href="http://www.pebinary.com" class="footerlogo" title="PE BINARY CO., LTD."><?php include "assets/include/logo.php"; ?></a>All rights reserved.<br>
-        <nav class="footer">
-          <ul class="list-inline" itemscope itemtype="http://schema.org/SiteNavigationElement">
-            <li><a href="http://www.pebinary.com/about/privacy.html" title="Privacy policy" itemprop="url"><i class="pe-lock"></i> <span class="hidden-md hidden-lg" itemprop="name">Privacy</span><span class="hidden-xs hidden-sm" itemprop="name">Privacy policy</span></a>&nbsp;</li>
-            <li><a href="http://www.pebinary.com/about/tos.html" title="Terms of services" itemprop="url"><i class="pe-gavel"></i> <span class="hidden-md hidden-lg" itemprop="name">TOS</span><span class="hidden-xs hidden-sm" itemprop="name">Terms of services</span></a>&nbsp;</li>
-            <li><a href="http://www.pebinary.com/about/terms.html" title="Terms &amp; conditions" itemprop="url"><i class="pe-balance-scale"></i> <span class="hidden-md hidden-lg" itemprop="name">Terms</span><span class="hidden-xs hidden-sm" itemprop="name">Terms &amp; conditions</span></a>&nbsp;</li>
-            <li><a href="http://www.pebinary.com/" title="Back to the main website" itemprop="url"><i class="pe-arrow-circle-left"></i> <span class="hidden-md hidden-lg" itemprop="name">The main website</span><span class="hidden-xs hidden-sm" itemprop="name">Back to the main website</span></a>&nbsp;</li>
-          </ul>
-        </nav>
-      </div>
-    </div>
-  </div>
-</footer>
-<form name="debug"><input class="btn btn-xs btn-block btn-danger" type="button" value="debug" onClick="debugWindow()"></form>
-<?php include "assets/include/js.php"; ?>
-</body>
-</html>
