--- admin/assets/include/header.php
@@ -1,38 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-<?php include "assets/include/meta.php"; ?>
-</head>
-<body>
-
-<header>
-  <div class="row">
-    <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12 header-1">
-      <div class="container">
-        <h1><?php include "assets/include/logo.php"; ?></h1>
-      </div>
-    </div>
-    <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12 header-2">
-      <nav class="navbar navbar-default navbar-static-top">
-        <div class="container">
-          <div class="navbar-header">
-            <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar">
-              <span class="sr-only">Toggle navigation</span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-            </button>
-            <a class="navbar-brand" href="/">Client zone</a>
-          </div>
-          <div id="navbar" class="navbar-collapse collapse">
-            <ul class="nav navbar-nav navbar-right">
-              <li><a href="http://www.siamsquare.org/admin"><i class="pe-power-off pe-lg"></i>&nbsp; Log-in</a></li>
-              <li><a href="/admin/contact.php"><i class="pe-envelope-o pe-lg"></i>&nbsp; Contact</a></li>
-              <li><a href="/admin/help.php"><i class="pe-question pe-lg"></i>&nbsp; Help</a></li>
-            </ul>
-          </div>
-        </div>
-      </nav>
-    </div>
-  </div>
-</header>
