--- admin/assets/include/logo.php
@@ -1,1 +1,0 @@
-<span class="logo1"><i class="pe-logo"></i></span> <span class="logo2">pe</span><span class="logo3">binary</span>
